Remove kotlin jvm testFixtures and dup test code
