+++ bindings/bdk-kotlin/android/build.gradle
@@ -25,12 +25,6 @@ android {
 }
 
 dependencies {
-   implementation(project(':jvm')) {
-       exclude group: 'net.java.dev.jna', module: 'jna'
-   }
-   api(project(':jvm')) {
-       exclude group: 'net.java.dev.jna', module: 'jna'
-   }
     implementation 'net.java.dev.jna:jna:5.8.0@aar'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'androidx.appcompat:appcompat:1.3.0'
@@ -38,12 +32,6 @@ dependencies {
     api "org.slf4j:slf4j-api:1.7.30"
 
     androidTestImplementation 'com.github.tony19:logback-android:2.0.0'
-   androidTestImplementation(testFixtures(project(':jvm'))) {
-       exclude group: 'net.java.dev.jna', module: 'jna'
-       exclude group: 'ch.qos.logback', module: 'logback-core'
-       exclude group: 'ch.qos.logback', module: 'logback-classic'
-       exclude group: 'junit', module: 'junit'
-   }
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
     androidTestImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.1'
