+++ bindings/bdk-kotlin/build.gradle
@@ -58,7 +58,3 @@ allprojects {
         }
     }
 }
-
-//task clean(type: Delete) {
-//  delete rootProject.buildDir
-//}
