+++ bindings/bdk-kotlin/jvm/build.gradle
@@ -25,9 +25,9 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation "net.java.dev.jna:jna:5.8.0"
     api "org.slf4j:slf4j-api:1.7.30"
-   testFixturesImplementation "junit:junit:4.13.2"
-   testFixturesImplementation "ch.qos.logback:logback-classic:1.2.3"
-   testFixturesImplementation "ch.qos.logback:logback-core:1.2.3"
+    testImplementation "junit:junit:4.13.2"
+    testImplementation "ch.qos.logback:logback-classic:1.2.3"
+    testImplementation "ch.qos.logback:logback-core:1.2.3"
 }
 
 afterEvaluate {
